[#44] :bug: Listing entity와 ListingDetail의 수정
- numBathroom가 소수점 값을 포함해야 하기 때문에 double 타입으로 수정.
- imageUrls는 List 자체로 빌드하기 때문에 @Singular 불필요.
- 불변객체이기 때문에 @Value로 @Data 대체.

+++ be/src/main/java/io/codesquad/group6/mockbnb/api/response/ListingDetail.java
@@ -20,7 +20,7 @@ public class ListingDetail {
     int capacity;
     int numBedrooms;
     int numBeds;
-    int numBathrooms;
+    double numBathrooms;
     double price;
     double rating;
     int numReviews;
